Move python tests to python/ray/tests/.
