+++ ci/jenkins_tests/multi_node_tests/many_drivers_test.py
@@ -6,8 +6,8 @@ import os
 import time
 
 import ray
-from ray.test.test_utils import (_wait_for_nodes_to_join, _broadcast_event,
-                                 _wait_for_event)
+from ray.tests.utils import (_wait_for_nodes_to_join, _broadcast_event,
+                             _wait_for_event)
 
 # This test should be run with 5 nodes, which have 0, 0, 5, 6, and 50 GPUs for
 # a total of 61 GPUs. It should be run with a large number of drivers (e.g.,
