+++ ci/stress_tests/application_cluster_template.yaml
