+++ ci/stress_tests/run_application_stress_tests.sh
