+++ ci/stress_tests/stress_testing_config.yaml
