+++ ci/travis/install-bazel.sh
