+++ ci/travis/install.sh
