+++ python/ray/rllib/train.py
@@ -8,7 +8,7 @@ import argparse
 import yaml
 
 import ray
-from ray.test.cluster_utils import Cluster
+from ray.tests.cluster_utils import Cluster
 from ray.tune.config_parser import make_parser
 from ray.tune.trial import resources_to_json
 from ray.tune.tune import _make_scheduler, run_experiments
