+++ python/ray/tests/test_array.py
@@ -10,7 +10,7 @@ import sys
 import ray
 import ray.experimental.array.remote as ra
 import ray.experimental.array.distributed as da
-import ray.test.cluster_utils
+import ray.tests.cluster_utils
 
 if sys.version_info >= (3, 0):
     from importlib import reload
@@ -75,7 +75,7 @@ def ray_start_two_nodes():
     ]:
         reload(module)
     # Start the Ray processes.
-    cluster = ray.test.cluster_utils.Cluster()
+    cluster = ray.tests.cluster_utils.Cluster()
     for _ in range(2):
         cluster.add_node(num_cpus=10)
     ray.init(redis_address=cluster.redis_address)
