+++ python/ray/tests/test_global_state.py
@@ -11,7 +11,7 @@ except ImportError:
 import time
 
 import ray
-from ray.test.cluster_utils import Cluster
+from ray.tests.cluster_utils import Cluster
 
 
 @pytest.fixture
